v0.7: working prototype
Changes to be committed:
	modified:   prototyping.ipynb
	modified:   qplib/__pycache__/qlang.cpython-313.pyc
	modified:   qplib/data/symbols.csv
	modified:   qplib/data/symbols.xlsx
	deleted:    qplib/qlang.1.py
	modified:   qplib/qlang.py
	renamed:    qplib/qlang.2.py -> qplib/qlang_old.py
